added beacon logic; 1. need to finish implementation of objective interaction, 2. need to add beacon/objective logic to networking

+++ common/game/Evolve.h
@@ -6,12 +6,12 @@
 
 class Evolve : public Objective {
     private:
-        int evo_amount; // amount to evolve
+        float evo_amount; // amount to evolve
 
     public:
         Evolve(GridPosition position, Restriction r = R_MONSTER);
 
-        int getEvoAmount();
+        float getEvoAmount();
 };
 
 #endif
